# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/doctor/add": {
            "post": {
                "description": "Return Add Doctor Successful",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctors"
                ],
                "summary": "Add Doctor",
                "parameters": [
                    {
                        "description": "Add a New Doctor",
                        "name": "doctor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.HandlerDoctor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseDataType"
                        }
                    }
                }
            }
        },
        "/api/v1/doctor/detail": {
            "get": {
                "description": "Return Detail Information of Doctors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctors"
                ],
                "summary": "Get Detail Information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Doctor Id",
                        "name": "doctorId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.HandlerDoctor"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ResponseDataType": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "types.HandlerDoctor": {
            "type": "object",
            "properties": {
                "clinicName": {
                    "type": "string"
                },
                "degree": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "doctorId": {
                    "type": "integer"
                },
                "doctorName": {
                    "type": "string"
                },
                "doctorSum": {
                    "type": "string"
                },
                "eduLocation": {
                    "type": "string"
                },
                "expYear": {
                    "type": "integer"
                },
                "fileUrl": {
                    "type": "string"
                },
                "language": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "workLocation": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}